refactor: Move ListSection to framework, so it can be reused
