+++ src/main/java/org/opentripplanner/routing/algorithm/mapping/RouteRequestToFilterChainMapper.java
@@ -7,10 +7,10 @@ import java.util.function.Consumer;
 import org.opentripplanner.ext.fares.FaresFilter;
 import org.opentripplanner.ext.ridehailing.RideHailingFilter;
 import org.opentripplanner.ext.stopconsolidation.ConsolidatedStopNameFilter;
+import org.opentripplanner.framework.collection.ListSection;
 import org.opentripplanner.routing.algorithm.filterchain.GroupBySimilarity;
 import org.opentripplanner.routing.algorithm.filterchain.ItineraryListFilterChain;
 import org.opentripplanner.routing.algorithm.filterchain.ItineraryListFilterChainBuilder;
-import org.opentripplanner.routing.algorithm.filterchain.ListSection;
 import org.opentripplanner.routing.algorithm.filterchain.deletionflagger.NumItinerariesFilterResults;
 import org.opentripplanner.routing.api.request.RouteRequest;
 import org.opentripplanner.routing.api.request.StreetMode;
